🎉  Feature/work on graphql resolvers (#17)
 ✨  Search Method added to crud controllers
 🎉  All resolvers works

+++ backend/controllers/crudController/crudMethods.ts
@@ -1,28 +1,16 @@
 import { Model } from 'mongoose';
 
-interface edgesType<T> {
-  edges: T[] | [];
-  pagination: {
-    page: number;
-    pages: number;
-    count: number;
-  };
-}
-
-interface crudMethodsType<T> {
-  read: (args: any) => Promise<T | null>;
-  create: (args: any) => Promise<T | null>;
-  update: (args: any) => Promise<T | null>;
-  delete: (args: any) => Promise<T | null>;
-  list: (args: any) => Promise<edgesType<T>>;
-}
+import { crudControllerType, listArgsType } from './crudControllerType';
 
 function crudMethods<T>(ModelEntity: Model<T>) {
-  let methods: crudMethodsType<T> = {
-    read: async (args) => {
+  let methods: crudControllerType<T> = {
+    create: async (args) => {
       try {
-        // Find document by id
-        const result = await ModelEntity.findOne({ _id: args?._id });
+        // Creating a new document in the collection
+        if (!args?.body) {
+          return null;
+        }
+        const result = await ModelEntity.create(args?.body);
         if (!result) {
           return null;
         } else {
@@ -33,10 +21,13 @@ function crudMethods<T>(ModelEntity: Model<T>) {
       }
     },
 
-    create: async (args) => {
+    read: async (args) => {
       try {
-        // Creating a new document in the collection
-        const result = await ModelEntity.create(args?.body);
+        // Find document by id
+        if (!args?._id) {
+          return null;
+        }
+        const result = await ModelEntity.findOne({ _id: args?._id });
         if (!result) {
           return null;
         } else {
@@ -49,6 +40,9 @@ function crudMethods<T>(ModelEntity: Model<T>) {
 
     update: async (args) => {
       try {
+        if (!args?._id || !args?.body) {
+          return null;
+        }
         // Find document by id and updates with the argsuired fields
         const result = await ModelEntity.findOneAndUpdate({ _id: args?._id }, args?.body, {
           new: true, // return the new result instead of the old one
@@ -66,6 +60,9 @@ function crudMethods<T>(ModelEntity: Model<T>) {
 
     delete: async (args) => {
       try {
+        if (!args?._id) {
+          return null;
+        }
         // Find the document by id and delete it
         const result = await ModelEntity.findOneAndDelete({ _id: args?._id }).exec();
         if (!result) {
@@ -79,18 +76,23 @@ function crudMethods<T>(ModelEntity: Model<T>) {
     },
 
     list: async (args) => {
-      const page = parseInt(args?.page) || 1;
-      const limit = parseInt(args?.limit) || 10;
-      const skip = page * limit - limit;
       try {
+        const page = parseInt(`${args?.page}`) || 1;
+        const limit = parseInt(`${args?.limit}`) || 10;
+        const sortBy = args?.sortBy || 'created';
+        const sortOrder: listArgsType['sortOrder'] = args?.sortOrder || 'desc';
+
+        const skip = page * limit - limit;
         //  Query the database for a list of all results
-        const resultsPromise = ModelEntity.find().skip(skip).limit(limit).sort({ created: 'desc' });
+        const resultsPromise = ModelEntity.find()
+          .skip(skip)
+          .limit(limit)
+          .sort({ [sortBy]: sortOrder });
         // Counting the total documents
         const countPromise = ModelEntity.count();
         // Resolving both promises
         const [result, count] = await Promise.all([resultsPromise, countPromise]);
-        console.log('🚀 ~ file: crudMethods.js ~ line 71 ~ list: ~ result', result);
-        console.log('🚀 ~ file: crudMethods.js ~ line 71 ~ list: ~ count', count);
+
         // Calculating total pages
         const pages = Math.ceil(count / limit);
         // Getting Pagination Object
@@ -116,6 +118,34 @@ function crudMethods<T>(ModelEntity: Model<T>) {
             count: 0,
           },
         };
+      }
+    },
+
+    search: async (args) => {
+      try {
+        if (args?.query === undefined || args?.query.trim() === '') {
+          return [];
+        }
+
+        const limit = parseInt(`${args?.limit}`) || 10;
+
+        const fieldsArray = args?.fields?.split(',') || [];
+
+        const fields: any = { $or: [] };
+
+        for (const field of fieldsArray) {
+          fields.$or.push({ [field]: { $regex: new RegExp(args?.query, 'i') } });
+        }
+
+        let results = await ModelEntity.find(fields).limit(limit);
+
+        if (results.length >= 1) {
+          return results;
+        } else {
+          return [];
+        }
+      } catch {
+        return [];
       }
     },
   };
